+++ example/src/app/app.module.ts
@@ -3,8 +3,8 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
-import { FontPickerModule, FontPickerConfigInterface } from 'angular2-font-picker';
-import { PerfectScrollbarModule, PerfectScrollbarConfigInterface } from 'angular2-perfect-scrollbar';
+import { FontPickerModule, FontPickerConfigInterface } from 'ngx-font-picker';
+import { PerfectScrollbarModule, PerfectScrollbarConfigInterface } from 'ngx-perfect-scrollbar';
 
 import { AppComponent } from './app.component';