+++ src/index.ts
@@ -1,1 +1,10 @@
-export * from './lib/index';
+export {
+  FontPickerComponent, FontPickerDirective, FontPickerService,
+  Font, GoogleFonts,
+  FontInterface, GoogleFontInterface, 
+  FontPickerConfig, FontPickerConfigInterface,
+  FontSizePipe, FontStylesPipe, StatefulSlicePipe,
+  provideDefaultConfig, provideForRootGuard,
+  FONT_PICKER_CONFIG, FONT_PICKER_GUARD,
+  FontPickerModule
+} from './lib/index';