+++ src/lib/index.ts
@@ -1,6 +1,21 @@
-export * from './font-picker.component';
-export * from './font-picker.directive';
-export * from './font-picker.module';
-export * from './font-picker.service';
-export * from './interfaces';
-export * from './pipes';
+export { FontPickerComponent } from './font-picker.component';
+
+export { FontPickerDirective } from './font-picker.directive';
+
+export { FontPickerService } from './font-picker.service';
+
+export {
+  provideDefaultConfig, provideForRootGuard,
+  FONT_PICKER_CONFIG, FONT_PICKER_GUARD,
+  FontPickerModule
+} from './font-picker.module';
+
+export {
+  Font, GoogleFonts,
+  FontInterface, GoogleFontInterface, 
+  FontPickerConfig, FontPickerConfigInterface
+} from './interfaces';
+
+export {
+  FontSizePipe, FontStylesPipe, StatefulSlicePipe
+} from './pipes';